src/weather-ingest-container-app/requirement.txt
requests
azure-identity
azure-eventhub
azure-keyvault-secrets
